Lots of improvements.

+++ src/test/java/com/wesleyhome/poi/api/assertions/WorkbookAssert.java
@@ -1,18 +1,16 @@
 package com.wesleyhome.poi.api.assertions;
 
-import org.apache.poi.openxml4j.exceptions.InvalidFormatException;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
 import org.assertj.core.api.AbstractAssert;
-import org.assertj.core.api.Assert;
 
+import java.awt.*;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Objects;
 import java.util.function.Supplier;
 
 import static java.util.Objects.requireNonNull;
@@ -22,40 +20,42 @@ public class WorkbookAssert extends AbstractAssert<WorkbookAssert, Workbook> {
         super(actual, WorkbookAssert.class);
     }
 
-    public static WorkbookAssert assertThat(Supplier<Workbook> supplier){
+    public static WorkbookAssert assertThat(Supplier<Workbook> supplier) {
         return assertThat(supplier.get());
     }
 
-    public static WorkbookAssert assertThat(Workbook workbook){
-        Path tempDir = null;
+    public static WorkbookAssert assertThat(Workbook workbook) {
         Path workbookPath = null;
-        try{
-            tempDir = Files.createTempDirectory("junit5-workbook");
-            workbookPath = tempDir.resolve("workbook.xlsx");
-            try(OutputStream os = Files.newOutputStream(workbookPath)){
+        try {
+            workbookPath = Files.createTempFile("junit5", ".xlsx");
+            System.out.println(workbookPath.toAbsolutePath());
+            try (OutputStream os = Files.newOutputStream(workbookPath)) {
                 workbook.write(os);
             }
             try (InputStream is = Files.newInputStream(workbookPath)) {
                 Workbook readWorkbook = WorkbookFactory.create(is);
                 return new WorkbookAssert(readWorkbook);
             }
-        } catch (IOException | InvalidFormatException e) {
+        } catch (IOException e) {
             throw new RuntimeException(e);
         } finally {
             try {
-                Files.deleteIfExists(requireNonNull(workbookPath));
-                Files.deleteIfExists(tempDir);
+                if (Desktop.isDesktopSupported()) {
+                    Desktop.getDesktop().open(workbookPath.toFile());
+                } else {
+                    Files.deleteIfExists(requireNonNull(workbookPath));
+                }
             } catch (IOException e) {
-                throw new RuntimeException(e);
+                // Don't fail the test because of this
+                e.printStackTrace();
             }
-
         }
     }
 
     public SheetAssert hasSheet(String sheetName) {
         isNotNull();
         Sheet sheet = actual.getSheet(sheetName);
-        if(sheet == null){
+        if (sheet == null) {
             failWithMessage("Expecting workbook to contain sheet <%s> but doesn't.", sheetName);
         }
         return new SheetAssert(sheet);
